As stated in branch title, working towards Validating input information between database and user input. Working torwards making next sign up form presentable.

+++ LotusTransformation/Models/ClientModels/ClientWorkInformation.cs
@@ -7,7 +7,7 @@ namespace LotusTransformation.Models
     {
         [Required]
         [Key]
-        public int Key { get; set; }
+        public long Key { get; set; }
 
         public string Occupation { get; set; }
